+++ example/decorators.py
@@ -4,43 +4,15 @@ Decorators
 
 # Standard Library
 import time
-from functools import wraps
 
 # Alliance Auth
 from allianceauth.services.hooks import get_extension_logger
 
-# Alliance Auth (External Libs)
-from app_utils.esi import EsiDailyDowntime, fetch_esi_status
-from app_utils.logging import LoggerAddTag
-
 # AA Example
 from example import __title__
-from example.app_settings import IS_TESTING
+from example.providers import AppLogger
 
-logger = LoggerAddTag(get_extension_logger(__name__), __title__)
-
-
-def when_esi_is_available(func):
-    """Make sure the decorated task only runs when esi is available.
-
-    Raise exception when ESI is offline.
-    Complete the task without running it when downtime is detected.
-
-    Automatically disabled during tests.
-    """
-
-    @wraps(func)
-    def outer(*args, **kwargs):
-        if IS_TESTING is not True:
-            try:
-                fetch_esi_status().raise_for_status()
-            except EsiDailyDowntime:
-                logger.info("Daily Downtime detected. Aborting.")
-                return None  # function will not run
-
-        return func(*args, **kwargs)
-
-    return outer
+logger = AppLogger(get_extension_logger(__name__), __title__)
 
 
 def log_timing(logs):